backend/requirements.txt
flask==2.3.2
flask-cors==3.0.10
azure-cosmos==4.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
azure-openai==1.0.0b1
azure-core==1.34.0
azure-identity==1.23.0